Bridgefy SDK 2.0.0
To redesign our security, we decided to think outside the box and consider the possibility of rewriting our entire model from scratch to build a new one that would get rid of all the reported issues right off the bat. Trying not to reinvent the wheel, however, we searched for an existing solution that we could use that was already validated by the industry and security experts. This is when we selected the Signal Protocol.